tests.pdf,.image - move pdf to image tests
