+++ tests/test_pdf.py
@@ -4,6 +4,10 @@ from acacore.siegfried import Siegfried
 
 from convertool.converters.base import dummy_base_file
 from convertool.converters.converter_pdf import ConverterPDF
+from convertool.converters.converter_pdf import ConverterPDFToImage
+from convertool.converters.converter_pdf import ConverterPDFToJPEG2000
+
+from .test_image import MIMETYPES
 
 
 # noinspection DuplicatedCode
@@ -19,3 +23,27 @@ def test_pdf_to_pdfa(test_files: dict[str, Path], output_dir: Path, siegfried: S
         assert output_files[0].is_file()
         match = siegfried.identify(output_files[0]).files[0]
         assert match.best_match().mime == "application/pdf"
+
+
+# noinspection DuplicatedCode
+def test_pdf_to_img(test_files: dict[str, Path], output_dir: Path, siegfried: Siegfried):
+    file = dummy_base_file(test_files["pdf-to-img.pdf"], test_files["pdf-to-img.pdf"].parent)
+    converter = ConverterPDFToImage(file)
+
+    for output in converter.outputs:
+        print(output)
+        output_files = converter.convert(output_dir, output)
+        assert len(output_files) >= 1
+        assert all(sf.best_match().mime == MIMETYPES[output] for sf in siegfried.identify(*output_files).files)
+
+
+# noinspection DuplicatedCode
+def test_pdf_to_jp2(test_files: dict[str, Path], output_dir: Path, siegfried: Siegfried):
+    file = dummy_base_file(test_files["pdf-to-img.pdf"], test_files["pdf-to-img.pdf"].parent)
+    converter = ConverterPDFToJPEG2000(file)
+
+    for output in converter.outputs:
+        print(output)
+        output_files = converter.convert(output_dir, output)
+        assert len(output_files) >= 1
+        assert all(sf.best_match().mime == MIMETYPES[output] for sf in siegfried.identify(*output_files).files)
